⚡️ ColorをintからUint8に

+++ minecraft.go
@@ -17,7 +17,7 @@ type Command func(rgb Color, x, y, z float64, blockId string) string
 
 // 0..255 RGB color
 type Color struct {
-	r, g, b int
+	r, g, b uint8
 }
 
 func scanBlockModel() (blockModelList map[string]string) {
@@ -141,9 +141,9 @@ func blockFilter(blockModelList map[string]string) (blockColor map[string]Color)
 			}
 
 			blockColor[blockID] = Color{
-				r: red / pixel,
-				g: green / pixel,
-				b: blue / pixel,
+				r: uint8(red / pixel),
+				g: uint8(green / pixel),
+				b: uint8(blue / pixel),
 			}
 		}
 	}
